+++ src/js/index.js
@@ -4,37 +4,10 @@ import ReactDOM from "react-dom";
 
 // include your styles into the webpack bundle
 import "../styles/index.css";
-import propTypes from "prop-types";
+import SimpleCounter from "./component/SimpleCounter";
 
-
-function SimpleCounter(props) {
-    return (
-        <div className="bigCounter">
-            <div className="calendar">
-                <i className="far fa-clock"></i>
-            </div>
-            <div className="six">{props.digitSix % 10}</div>
-            <div className="five">{props.digitFive % 10}</div>
-            <div className="four">{props.digitFour % 10}</div>
-            <div className="three">{props.digitThree % 10}</div>
-            <div className="two">{props.digitTwo % 10}</div>
-            <div className="one">{props.digitOne % 10}</div>
-        </div>
-    )
-}
-
-SimpleCounter.propTypes = {
-    digitSix: propTypes.number,
-    digitFive: propTypes.number,
-    digitFour: propTypes.number,
-    digitThree: propTypes.number,
-    digitTwo: propTypes.number,
-    digitOne: propTypes.number
-}
 
 let counter = 0;
-
-
 setInterval(function () {
     const six = Math.floor(counter / 100000);
     const five = Math.floor(counter / 10000);
@@ -42,9 +15,8 @@ setInterval(function () {
     const three = Math.floor(counter / 100);
     const two = Math.floor(counter / 10);
     const one = Math.floor(counter / 1);
-    console.log(six, five, four, three, two, one);
-    counter++;
     ReactDOM.render(<SimpleCounter digitOne={one} digitTwo={two} digitThree={three} digitFour={four} digitFive={five} digitSix={six} />,
         document.querySelector("#app"));
+    counter++;
 }, 1000)
